Add ShowK TypeClass
This is a typeclass similar to Show as SemigroupK is to Semigroup.

It is particularly useful in defining a `Show` (not `ShowK`) instance for GADT which has a high number of differing type parameters, all wrapped in the same `F[_]`.
For example, given this case class.

```scala
final case class Foo[F[_]](a: F[Int], b: F[Long], c: F[String], d: F[Short], e: F[Byte])
```

In order to write a `Show` instance for it, one would currently have to demand a `Show` instance for each distinct type in the product.

```scala
object Foo {
   implicit def show[F[_]](implicit A: Show[F[Int]], B: Show[F[Long]], C: Show[F[String]], D: Show[F[Short]], e: Show[F[Byte]]): Show[Foo[F]] = {
    new Show[Foo[F]] {
      override show(t: Foo[F]): String =
        show"Foo(${t.a}, ${t.b}, ${t.c}, ${t.d}, ${t.e})
    }
  }
}
```

This can be quite cumbersome. However with `ShowK` one can merely do this.

```scala
  implicit def show[F[_]: ShowK]: Show[Foo[F]] = {
    new Show[Foo[F]] {
      override show(t: Foo[F]): String =
        showK"Foo(${t.a}, ${t.b}, ${t.c}, ${t.d}, ${t.e})
    }
  }
```

Demanding that only the `F[_]` have an instance of `ShowK`.

This commit also adds a new `StringContext` `showK` which can be used to interpolate values which have _either_ a `ShowK` instance _or_ a `Show` instance.